Cardano.Launcher

Description

This module contains a mechanism for launching external processes, ensuring that they are terminated on exceptions.

Documentation

data Command

Represent a command to execute. Args are provided as a list where options are expected to be prefixed with `--` or -. For example:

Command "cardano-wallet"
    [ "server"
    , "--port", "8080"
    , "--network", "mainnet"
    ] (return ())
    Inherit
    Inherit

Constructors

Command

Fields

cmdName :: String

cmdArgs :: [String]

cmdSetup :: IO ()

An extra action to run _before_ the command

cmdInput :: StdStream

Input to supply to command

cmdOutput :: StdStream

What to do with stdout & stderr

data StdStream

Constructors

Inherit
Inherit Handle from parent

UseHandle Handle
Use the supplied Handle

CreatePipe
Create a new pipe. The returned Handle will use the default encoding and newline translation mode (just like Handles created by openFile).

NoStream
Close the stream's file descriptor without passing a Handle. On POSIX systems this may lead to strange behavior in the child process because attempting to read or write after the file has been closed throws an error. This should only be used with child processes that don't use the file descriptor at all. If you wish to ignore the child process's output you should either create a pipe and drain it manually or pass a Handle that writes to /dev/null.

data ProcessHasExited

ProcessHasExited is used by a monitoring thread to signal that the process has exited.

Constructors

ProcessDidNotStart String IOException

ProcessHasExited String ExitCode

withBackendProcess

Arguments

:: MonadUnliftIO m

=> Tracer m LauncherLog
Logging

-> Command
Command description

-> (Maybe Handle -> ProcessHandle -> m a)
Action to execute while process is running.

-> m (Either ProcessHasExited a)

Starts a command in the background and then runs an action. If the action finishes (through an exception or otherwise) then the process is terminated (see withCreateProcess) for details. If the process exits, the action is cancelled. The return type reflects those two cases.

The action receives the ProcessHandle and stdin Handle as arguments.

withBackendCreateProcess

Arguments

:: forall m a. MonadUnliftIO m

=> Tracer m LauncherLog
Logging

-> CreateProcess
Command description

-> (Maybe Handle -> ProcessHandle -> m a)
Action to execute while process is running.

-> m (Either ProcessHasExited a)

A variant of withBackendProcess which accepts a general CreateProcess object. This version also has nicer async properties than withCreateProcess.

This function should ensure:

If the action finishes or throws an exception, then the process is also terminated.

After the process is sent the signal to terminate, this function will block until the process has actually exited - unless that takes longer than the 5 second timeout. After the timeout has lapsed, the process will be sent a kill signal.

If the process exits, then the action is cancelled.

fixme: This is more or less a reimplementation of withProcessWait (except for wait timeout). The launcher code should be converted to use typed-process.

Logging

data LauncherLog

Constructors

MsgLauncherStart String [String]

WithProcessInfo String Text LaunchedProcessLog

MsgLauncherCleanup

MsgLauncherCleanupTimedOut Int

MsgLauncherCleanupFinished

MsgLauncherFinish (Maybe ProcessHasExited)