Cardano.Launcher.Node

Description

Provides a function to launch cardano-node.

Startup

withCardanoNode

Arguments

:: Tracer IO LauncherLog
Trace for subprocess control logging

-> CardanoNodeConfig

-> (CardanoNodeConn -> IO a)
Callback function with a socket filename and genesis params

-> IO (Either ProcessHasExited a)

Spawns a cardano-node process.

IMPORTANT: cardano-node must be available on the current path.

data CardanoNodeConfig

A subset of the cardano-node CLI parameters, used for starting the backend.

Constructors

CardanoNodeConfig

Fields

nodeDir :: FilePath

nodeConfigFile :: FilePath

nodeTopologyFile :: FilePath

nodeDatabaseDir :: FilePath

nodeDlgCertFile :: Maybe FilePath

nodeSignKeyFile :: Maybe FilePath

nodeOpCertFile :: Maybe FilePath

nodeKesKeyFile :: Maybe FilePath

nodeVrfKeyFile :: Maybe FilePath

nodePort :: Maybe NodePort

nodeLoggingHostname :: Maybe String

newtype NodePort

Constructors

NodePort

Fields

unNodePort :: Int

cardano-node Snockets

data CardanoNodeConn

Parameters for connecting to the node.

cardanoNodeConn :: FilePath -> Either String CardanoNodeConn

Produces a CardanoNodeConn if the socket path or pipe name (depending on isWindows) is valid.

nodeSocketFile :: CardanoNodeConn -> FilePath

Gets the socket filename or pipe name from CardanoNodeConn. Whether it's a unix socket or named pipe depends on the value of isWindows.

isWindows :: Bool