Cardano.Launcher

Description

This module contains a mechanism for launching external processes together, and provides the functionality needed to kill them all if one goes down. (would be achieved using monitor and kill in combination)

Documentation

data Command

Represent a command to execute. Args are provided as a list where options are expected to be prefixed with `--` or -. For example:

Command "cardano-wallet"
    [ "server"
    , "--port", "8080"
    , "--network", "mainnet"
    ] (return ())
    Inherit

Constructors

Command

Fields

cmdName :: String

cmdArgs :: [String]

cmdSetup :: IO ()

An extra action to run _before_ the command

cmdOutput :: StdStream

What to do with stdout & stderr

data StdStream

Constructors

Inherit
Inherit Handle from parent

UseHandle Handle
Use the supplied Handle

CreatePipe
Create a new pipe. The returned Handle will use the default encoding and newline translation mode (just like Handles created by openFile).

NoStream
Close the stream's file descriptor without passing a Handle. On POSIX systems this may lead to strange behavior in the child process because attempting to read or write after the file has been closed throws an error. This should only be used with child processes that don't use the file descriptor at all. If you wish to ignore the child process's output you should either create a pipe and drain it manually or pass a Handle that writes to /dev/null.

data ProcessHasExited

ProcessHasExited is used by a monitoring thread to signal that the process has exited.

Constructors

ProcessHasExited String ExitCode

launch :: [Command] -> IO ProcessHasExited

Run a bunch of command in separate processes. Note that, this operation is blocking and will throw when one of the given commands terminates. Commands are therefore expected to be daemon or long-running services.

installSignalHandlers :: IO ()

Convert any SIGTERM received to SIGINT, for which the runtime system has handlers that will correctly clean up sub-processes.