Cardano.Launcher

Description

This module contains a mechanism for launching external processes, ensuring that they are terminated on exceptions.

Documentation

data Command

Represent a command to execute. Args are provided as a list where options are expected to be prefixed with `--` or -. For example:

Command "cardano-wallet"
    [ "server"
    , "--port", "8080"
    , "--network", "mainnet"
    ] (return ())
    Inherit

Constructors

Command

Fields

cmdName :: String

cmdArgs :: [String]

cmdSetup :: IO ()

An extra action to run _before_ the command

cmdOutput :: StdStream

What to do with stdout & stderr

data StdStream

Constructors

Inherit
Inherit Handle from parent

UseHandle Handle
Use the supplied Handle

CreatePipe
Create a new pipe. The returned Handle will use the default encoding and newline translation mode (just like Handles created by openFile).

NoStream
Close the stream's file descriptor without passing a Handle. On POSIX systems this may lead to strange behavior in the child process because attempting to read or write after the file has been closed throws an error. This should only be used with child processes that don't use the file descriptor at all. If you wish to ignore the child process's output you should either create a pipe and drain it manually or pass a Handle that writes to /dev/null.

data ProcessHasExited

ProcessHasExited is used by a monitoring thread to signal that the process has exited.

Constructors

ProcessDidNotStart String IOException

ProcessHasExited String ExitCode

launch :: Trace IO LauncherLog -> [Command] -> IO ProcessHasExited

Run a bunch of command in separate processes. Note that, this operation is blocking and will throw when one of the given commands terminates. Commands are therefore expected to be daemon or long-running services.

withBackendProcess

Arguments

:: Trace IO LauncherLog
Logging

-> Command
Command description

-> IO a
Action to execute while process is running.

-> IO (Either ProcessHasExited a)

Starts a command in the background and then runs an action. If the action finishes (through an exception or otherwise) then the process is terminated (see withCreateProcess) for details. If the process exits, the action is cancelled. The return type reflects those two cases.

Program startup

installSignalHandlers :: Trace IO Text -> IO ()

Convert any SIGTERM received to SIGINT, for which the runtime system has handlers that will correctly clean up sub-processes.

setUtf8Encoding :: IO ()

Force the locale text encoding to UTF-8. This is needed because the CLI prints UTF-8 characters regardless of the LANG environment variable or any other settings.

On Windows the current console code page is changed to UTF-8.

Logging

data LauncherLog

Constructors

MsgLauncherStart Command

MsgLauncherStarted String Text

MsgLauncherFinish ProcessHasExited

MsgLauncherCleanup

transformLauncherTrace :: Trace IO Text -> Trace IO LauncherLog