Cardano.Startup

Description

This module contains functions relating to startup and shutdown of the cardano-wallet serve program.

Program startup

withUtf8Encoding :: IO a -> IO a

Force the locale text encoding to UTF-8. This is needed because the CLI prints UTF-8 characters regardless of the LANG environment variable or any other settings.

On Windows the current console code page is changed to UTF-8.

Clean shutdown

withShutdownHandler :: Tracer IO ShutdownHandlerLog -> IO a -> IO (Maybe a)

Runs the given action with a cross-platform clean shutdown handler.

This is necessary when running cardano-wallet as a subprocess of Daedalus. For more details, see https://github.com/input-output-hk/cardano-launcher/blob/master/docs/windows-clean-shutdown.md

It works simply by reading from stdin, which is otherwise unused by the API server. Once end-of-file is reached, it cancels the action, causing the program to shut down.

So, when running cardano-wallet as a subprocess, the parent process should pass a pipe for stdin, then close the pipe when it wants cardano-wallet to shut down.

TODO: may need to add forkIO if hGet blocks on windows

withShutdownHandler' :: Tracer IO ShutdownHandlerLog -> Handle -> IO a -> IO (Maybe a)

A variant of withShutdownHandler where the handle to read can be chosen.

installSignalHandlers :: IO () -> IO ()

Convert any SIGTERM received to SIGINT, for which the runtime system has handlers that will correctly clean up sub-processes.

Logging

data ShutdownHandlerLog

Constructors

MsgShutdownHandler Bool

MsgShutdownEOF

MsgShutdownError IOException